% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterchooser.R
\name{characterchooser}
\alias{characterchooser}
\title{characterchooser}
\usage{
characterchooser(my_fave = character)
}
\arguments{
\item{my_fave}{Your favorite StarWars character. Default = sampled from dplyr::starwars name column.}
}
\description{
This is a function that randomly selects Star Wars characters from the dplyr database and then states that they are the users favorite character.
}
\examples{
characterchooser()
}
\keyword{Star}
\keyword{Wars}
